Save as html with embedded UMD
